[refactor] changes folder structure of rest-server
