+++ rest-server/src/db/index.js
@@ -19,4 +19,4 @@ mongoose.connection.on('disconnected', () => {
 
 const db = mongoose.connection;
 
-module.exports = db;
+export default db;
