+++ rest-server/src/db/models/user.js
@@ -13,4 +13,4 @@ const userSchema = mongoose.Schema({
 
 const User = mongoose.model('User', userSchema);
 
-module.exports = { User };
+export default User;
