+++ rest-server/src/lib/bcrypt.js
